Fix compiler warning by switching loop var type to size_t

+++ photon/src/PhotonNetworkLayer.cpp
@@ -57,7 +57,7 @@ bool PhotonNetworkLayer::getPacket(uint8_t* buffer, size_t maxBufferLength) {
 
   size_t packet_length = min(_packetizer->packetLength(), maxBufferLength);
   // iterate through all bytes in RS485 object and plop them in the buffer
-  for(int i = 0; i<packet_length; i++){
+  for(size_t i = 0; i<packet_length; i++){
     buffer[i] = (*_bus)[i];
   }
